Changes for Release 2406.0.0
- Updated pnpm to version 9
- Updated `schema.json` to CoreMedia Content Cloud v12 - 2604.0
- Changed `.graphqlconfig` to new format `graphql.config.yml`
- Added `schema.graphql` into the workspace for code completion
- Added support for CoreMedia Engagement Cloud configuration from settings.
- Optimized stitching server (logging, code cleanup)
- Updated storybook to v8
- Updated minor versions of dependencies
