+++ .github/workflows/ci.yml
@@ -19,7 +19,7 @@ jobs:
     - name: Use pnpm
       uses: pnpm/action-setup@v3
       with:
-        version: 8
+        version: 9
 
     - name: Use Node.js
       uses: actions/setup-node@v4
