--- .graphqlconfig
@@ -1,15 +1,0 @@
-{
-  "name": "CoreMedia Headless Stitching Server Schema",
-  "schemaPath": "packages/graphql-layer/src/__downloaded__/schema.graphql",
-  "extensions": {
-    "endpoints": {
-      "GraphQL Endpoint": {
-        "url": "http://localhost:4000/graphql",
-        "headers": {
-          "user-agent": "JS GraphQL"
-        },
-        "introspect": true
-      }
-    }
-  }
-}
