+++ Dockerfile
@@ -1,2 +1,2 @@
 FROM node:20
-RUN npm install -g pnpm@8
+RUN npm install -g pnpm@9
