+++ apps/spark/.env.example
@@ -25,6 +25,6 @@ VITE_FQDN=https://<headless-stitching-server.example>
 # Set loglevel, default is "warn" for prod and "info" for dev.
 #VITE_LOGLEVEL=debug
 
-# Byside Integration, default URL: https://bywe2.byside.com/agent/bwc_we2.js
-#VITE_BYSIDE_WEBCARE_ID=<your-webcare-id>
-#VITE_BYSIDE_WEBCARE_SCRIPT_URL=<your-custom-tag-url>
+# Engagement Cloud Integration, default URL: https://bywe2.byside.com/agent/bwc_we2.js
+#VITE_ENGAGEMENT_CLOUD_ID=<your-webcare-id>
+#VITE_ENGAGEMENT_CLOUD_TAG_URL=<your-custom-tag-url>
