+++ apps/spark/package.json
@@ -17,25 +17,25 @@
     "build-storybook": "storybook build -o dist/storybook"
   },
   "dependencies": {
-    "@apollo/client": "^3.10.1",
+    "@apollo/client": "^3.10.5",
     "@coremedia-labs/graphql-layer": "workspace:^",
     "@coremedia-labs/preview-integration": "workspace:^",
     "@coremedia-labs/view-dispatcher": "workspace:^",
-    "@js-joda/core": "^5.6.2",
-    "@js-joda/timezone": "^2.21.0",
-    "country-flag-icons": "^1.5.11",
+    "@js-joda/core": "^5.6.3",
+    "@js-joda/timezone": "^2.21.1",
+    "country-flag-icons": "^1.5.12",
     "crypto-hash": "^2.0.1",
-    "graphql": "^16.8.1",
-    "i18next": "^23.11.2",
+    "graphql": "^16.8.2",
+    "i18next": "^23.11.5",
     "i18next-browser-languagedetector": "^7.2.1",
-    "i18next-http-backend": "^2.5.1",
+    "i18next-http-backend": "^2.5.2",
     "loglevel": "^1.9.1",
     "query-string": "^7.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet-async": "^1.3.0",
     "react-i18next": "^11.18.6",
-    "react-is": "^18.2.0",
+    "react-is": "^18.3.1",
     "react-player": "^2.16.0",
     "react-responsive": "^9.0.2",
     "react-router-dom": "^5.3.4",
@@ -45,37 +45,37 @@
     "styled-components": "^5.3.11"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^7.6.18",
-    "@storybook/addon-backgrounds": "^7.6.18",
-    "@storybook/addon-docs": "^7.6.18",
-    "@storybook/addon-essentials": "^7.6.18",
-    "@storybook/addon-links": "^7.6.18",
-    "@storybook/addon-measure": "^7.6.18",
-    "@storybook/addon-outline": "^7.6.18",
-    "@storybook/client-api": "^7.6.17",
-    "@storybook/client-logger": "^7.6.18",
-    "@storybook/node-logger": "^7.6.18",
-    "@storybook/react": "^7.6.18",
-    "@storybook/react-vite": "^7.6.18",
-    "@types/node": "^20.12.7",
+    "@chromatic-com/storybook": "^1.5.0",
+    "@storybook/addon-actions": "^8.1.10",
+    "@storybook/addon-backgrounds": "^8.1.10",
+    "@storybook/addon-docs": "^8.1.10",
+    "@storybook/addon-essentials": "^8.1.10",
+    "@storybook/addon-links": "^8.1.10",
+    "@storybook/addon-measure": "^8.1.10",
+    "@storybook/addon-outline": "^8.1.10",
+    "@storybook/client-logger": "^8.1.10",
+    "@storybook/node-logger": "^8.1.10",
+    "@storybook/react": "^8.1.10",
+    "@storybook/react-vite": "^8.1.10",
+    "@types/node": "^20.14.5",
     "@types/react": "^17.0.80",
     "@types/react-dom": "^17.0.25",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-slick": "^0.23.13",
     "@types/styled-components": "^5.1.34",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
-    "@vitejs/plugin-react-swc": "^3.6.0",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "@typescript-eslint/parser": "^7.13.1",
+    "@vitejs/plugin-react-swc": "^3.7.0",
     "dotenv": "^16.4.5",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react": "^7.34.3",
     "eslint-plugin-storybook": "^0.8.0",
-    "prettier": "^3.2.5",
-    "storybook": "^7.6.18",
+    "prettier": "^3.3.2",
+    "storybook": "^8.1.10",
     "typedoc": "^0.25.13",
     "typescript": "~5.2.2",
     "vite": "^4.5.3",
