+++ apps/spark/public/locales/fr/translation.json
@@ -63,6 +63,7 @@
     "type_CMArticle": "Stories",
     "type_CMCategory": "Categories",
     "type_CMChannel": "Pages",
+    "type_CMCollection": "Collections",
     "type_CMEvent": "Events",
     "type_CMExternalChannel": "External Channels",
     "type_CMExternalPage": "External Pages",
