+++ apps/standalone-fragment/package.json
@@ -9,18 +9,18 @@
     "lint:ci": "eslint \"src/**/*.+(ts|tsx)\""
   },
   "dependencies": {
-    "@apollo/client": "^3.10.1",
+    "@apollo/client": "^3.10.5",
     "@coremedia-labs/graphql-layer": "workspace:^",
-    "graphql": "^16.8.1",
+    "graphql": "^16.8.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.4",
+    "@babel/core": "^7.24.7",
     "@types/react": "^17.0.80",
     "@types/react-dom": "^17.0.25",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "@typescript-eslint/parser": "^7.13.1",
     "babel-loader": "^9.1.3",
     "babel-preset-react-app": "^10.0.1",
     "copy-webpack-plugin": "^12.0.2",
@@ -28,11 +28,11 @@
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.34.1",
-    "prettier": "^3.2.5",
+    "eslint-plugin-react": "^7.34.3",
+    "prettier": "^3.3.2",
     "style-loader": "^3.3.4",
     "typescript": "~5.2.2",
-    "webpack": "^5.91.0",
+    "webpack": "^5.92.0",
     "webpack-cli": "^5.1.4"
   },
   "browserslist": {
