+++ servers/mocking/package.json
@@ -18,9 +18,9 @@
     "express": "^4.19.2",
     "express-http-proxy": "^2.0.0",
     "globby": "^11.1.0",
-    "graphql": "^16.8.1",
+    "graphql": "^16.8.2",
     "import-fresh": "^3.3.0",
-    "lodash": "^4.5.0",
+    "lodash": "^4.17.21",
     "mime-types": "^2.1.35",
     "minimist": "^1.2.8",
     "morgan": "^1.10.0",
@@ -34,7 +34,7 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "jest": "^29.7.0",
-    "prettier": "^3.2.5",
+    "prettier": "^3.3.2",
     "supertest": "^7.0.0"
   },
   "jest": {
