+++ servers/stitching/package.json
@@ -11,12 +11,12 @@
   },
   "dependencies": {
     "@apollo/server": "^4.10.4",
-    "@graphql-tools/delegate": "^10.0.4",
+    "@graphql-tools/delegate": "^10.0.11",
     "@graphql-tools/load": "^8.0.2",
-    "@graphql-tools/schema": "^10.0.3",
-    "@graphql-tools/stitch": "^9.1.2",
+    "@graphql-tools/schema": "^10.0.4",
+    "@graphql-tools/stitch": "^9.2.9",
     "@graphql-tools/url-loader": "^8.0.2",
-    "@graphql-tools/utils": "^10.1.3",
+    "@graphql-tools/utils": "^10.2.2",
     "@graphql-tools/wrap": "^10.0.5",
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
@@ -24,7 +24,7 @@
     "express": "^4.19.2",
     "express-actuator": "^1.8.4",
     "express-http-proxy": "^2.0.0",
-    "graphql": "^16.8.1",
+    "graphql": "^16.8.2",
     "typescript": "~5.2.2",
     "winston": "^3.13.0"
   },
@@ -34,14 +34,14 @@
     "@types/express": "^4.17.21",
     "@types/express-actuator": "^1.8.3",
     "@types/express-http-proxy": "^1.6.6",
-    "@types/node": "^20.12.7",
-    "@typescript-eslint/eslint-plugin": "^7.7.1",
-    "@typescript-eslint/parser": "^7.7.1",
+    "@types/node": "^20.14.5",
+    "@typescript-eslint/eslint-plugin": "^7.13.1",
+    "@typescript-eslint/parser": "^7.13.1",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
-    "nodemon": "^3.1.0",
-    "prettier": "^3.2.5",
+    "nodemon": "^3.1.3",
+    "prettier": "^3.3.2",
     "ts-node": "^10.9.2"
   }
 }
