--- servers/stitching/src/types.d.ts
@@ -1,1 +1,0 @@
-/// <reference lib="dom" />
